update uint64_t to int for random seed

+++ include/gda_sa.h
@@ -18,7 +18,7 @@ class LISA;
 LISA* gda_localmoran(GeoDaWeight *w,
                         const std::vector<double> &data,
                         const std::vector<bool> &undefs = std::vector<bool>(),
-                     int nCPUs = 8, int permutations = 999, uint64_t last_seed_used = 123456789);
+                     int nCPUs = 8, int permutations = 999, int last_seed_used = 123456789);
 
 /**
  *
@@ -29,7 +29,7 @@ LISA* gda_localmoran(GeoDaWeight *w,
  */
 LISA* gda_geary(GeoDaWeight *w, const std::vector<double> &data,
                     const std::vector<bool> &undefs = std::vector<bool>(),
-                int nCPUs = 8, int permutations = 999, uint64_t last_seed_used = 123456789);
+                int nCPUs = 8, int permutations = 999, int last_seed_used = 123456789);
 
 /**
  *
@@ -40,7 +40,7 @@ LISA* gda_geary(GeoDaWeight *w, const std::vector<double> &data,
  */
 LISA* gda_multigeary(GeoDaWeight *w, const std::vector<std::vector<double> > &data,
                     const std::vector<std::vector<bool> > &undefs = std::vector<std::vector<bool> >(),
-                     int nCPUs = 8, int permutations = 999, uint64_t last_seed_used = 123456789);
+                     int nCPUs = 8, int permutations = 999, int last_seed_used = 123456789);
 
 /**
  *
@@ -51,11 +51,11 @@ LISA* gda_multigeary(GeoDaWeight *w, const std::vector<std::vector<double> > &da
  */
 LISA* gda_joincount(GeoDaWeight *w, const std::vector<double> &data,
                             const std::vector<bool> &undefs = std::vector<bool>(),
-                    int nCPUs = 8, int permutations = 999, uint64_t last_seed_used = 123456789);
+                    int nCPUs = 8, int permutations = 999, int last_seed_used = 123456789);
 
 LISA* gda_multijoincount(GeoDaWeight *w, const std::vector<std::vector<double> > &data,
                 const std::vector<std::vector<bool> > &undefs = std::vector<std::vector<bool> >(),
-                         int nCPUs = 8, int permutations = 999, uint64_t last_seed_used = 123456789);
+                         int nCPUs = 8, int permutations = 999, int last_seed_used = 123456789);
 /**
  *
  * @param w
@@ -65,7 +65,7 @@ LISA* gda_multijoincount(GeoDaWeight *w, const std::vector<std::vector<double> >
  */
 LISA* gda_localg(GeoDaWeight *w, const std::vector<double> &data,
                  const std::vector<bool> &undefs = std::vector<bool>(),
-                 int nCPUs = 8, int permutations = 999, uint64_t last_seed_used = 123456789);
+                 int nCPUs = 8, int permutations = 999, int last_seed_used = 123456789);
 
 /**
  *
@@ -76,7 +76,7 @@ LISA* gda_localg(GeoDaWeight *w, const std::vector<double> &data,
  */
 LISA* gda_localgstar(GeoDaWeight *w, const std::vector<double> &data,
                          const std::vector<bool> &undefs = std::vector<bool>(),
-                     int nCPUs = 8, int permutations = 999, uint64_t last_seed_used = 123456789);
+                     int nCPUs = 8, int permutations = 999, int last_seed_used = 123456789);
 
 /**
  *
@@ -88,7 +88,7 @@ LISA* gda_localgstar(GeoDaWeight *w, const std::vector<double> &data,
  */
 LISA* gda_quantilelisa(GeoDaWeight* w, unsigned int k,  unsigned int quantile, const std::vector<double> &data,
                        const std::vector<bool> &undefs = std::vector<bool>(),
-                       int nCPUs = 8, int permutations = 999, uint64_t last_seed_used = 123456789);
+                       int nCPUs = 8, int permutations = 999, int last_seed_used = 123456789);
 
 /**
  *
